+++ internal/shared/version/version.go
